Update README.md
The global mobile application market was valued at $106.27 billion in 2018, $154.05 billion in 2019 and is expected to grow at a compound annual growth rate (CAGR) of 11.5% from 2020 to 2027. 

Customer behavior must be captured so FinTech has released a mobile app with two versions: free and premium. So, in the free version, they provide a premium feature for 24 hours. We have 50000 rows (total number of survey participants) and 12 attributes (describe the behavior tendency of each participants)